converters.text - move text converters to their own file, add "text" to "txt" converter with "stripnull" option
